lib/fs/readlink/: readlinknul(): Fix return type
Fixes: 419ce14b6f72 (2024-11-01, "lib/fs/readlink/: readlinknul(): Add function")
Cc: Serge Halyn <serge@hallyn.com>

+++ lib/fs/readlink/readlinknul.c
@@ -7,7 +7,8 @@
 #include "fs/readlink/readlinknul.h"
 
 #include <stddef.h>
+#include <sys/types.h>
 
 
-extern inline int readlinknul(const char *restrict link, char *restrict buf,
+extern inline ssize_t readlinknul(const char *restrict link, char *restrict buf,
     size_t size);
